removed unused method, and fix bugs. Also recreating flow to be more efficient

+++ src/main/java/com/tokopakde/toryment/service/impl/CategoryServiceImpl.java
@@ -8,6 +8,7 @@ import com.tokopakde.toryment.dto.category.CategoryResDTO;
 import com.tokopakde.toryment.dto.category.CreateCategoryReqDTO;
 import com.tokopakde.toryment.dto.category.UpdateCategoryReqDTO;
 import com.tokopakde.toryment.dto.pagination.PageRes;
+import com.tokopakde.toryment.exceptiohandler.exception.ConflictException;
 import com.tokopakde.toryment.exceptiohandler.exception.DataIntegrationException;
 import com.tokopakde.toryment.exceptiohandler.exception.DuplicateException;
 import com.tokopakde.toryment.exceptiohandler.exception.NotFoundException;
@@ -15,6 +16,7 @@ import com.tokopakde.toryment.mapper.CategoryMapper;
 import com.tokopakde.toryment.mapper.PageMapper;
 import com.tokopakde.toryment.model.company.Category;
 import com.tokopakde.toryment.repository.CategoryRepo;
+import com.tokopakde.toryment.repository.ProductRepo;
 import com.tokopakde.toryment.service.BaseService;
 import com.tokopakde.toryment.service.CategoryService;
 import lombok.RequiredArgsConstructor;
@@ -26,6 +28,7 @@ import org.springframework.stereotype.Service;
 @Service
 public class CategoryServiceImpl extends BaseService implements CategoryService{
     private final CategoryRepo categoryRepo;
+    private final ProductRepo productRepo;
     private final CategoryMapper mapper;
     private final PageMapper pageMapper;
 
@@ -44,7 +47,7 @@ public class CategoryServiceImpl extends BaseService implements CategoryService{
     @Override
     public CreateResDTO createCategory(CreateCategoryReqDTO request) {
         if (categoryRepo.existsByCode(request.getCode())) {
-            throw new DuplicateException("Code Is Not Available");
+            throw new DuplicateException("This Code Is Used By Another Category");
         }
 
         var category = mapper.mapToEntity(request);
@@ -60,7 +63,10 @@ public class CategoryServiceImpl extends BaseService implements CategoryService{
             throw new DataIntegrationException("Error Updating Category, Please Refresh The Page");
         }
 
-        validateUniqueCode(category.getCode(), request.getCode(), categoryRepo::findByCode);
+        if (!category.getCode().equals(request.getCode())
+                && categoryRepo.existsByCode(request.getCode())) {
+            throw new DuplicateException("This Code Is Used By Another Category");
+        }
 
         category.setCode(request.getCode());
         category.setName(request.getName());
@@ -71,6 +77,11 @@ public class CategoryServiceImpl extends BaseService implements CategoryService{
     @Override
     public CommonResDTO deleteCategory(String id) {
         var category = findCategoryById(id);
+
+        if (productRepo.existsByCategory(category)) {
+            throw new ConflictException("Category Cannot Be Deleted, Because It Is Used By Existing Product");
+        }
+
         categoryRepo.delete(category);
         return new CommonResDTO(Message.DELETED.getDescription());
     }
